REFACTOR: Move source selection function to label module
The select_sources function can be used in many situation, not only for
simulation.

This commit moves the select_sources function to the label module and
adds to the mne name space. The tests are also moved.
